Add files for _detect_program_name, _expand_args, and make_default_short_help.
