--- tests/test_utils.py
@@ -1,79 +1,0 @@
-import os
-import pathlib
-
-import pytest
-
-import click._termui_impl
-import click.utils
-
-
-class MockMain:
-    __slots__ = "__package__"
-
-    def __init__(self, package_name):
-        self.__package__ = package_name
-
-
-@pytest.mark.parametrize(
-    ("path", "main", "expected"),
-    [
-        ("example.py", None, "example.py"),
-        (str(pathlib.Path("/foo/bar/example.py")), None, "example.py"),
-        ("example", None, "example"),
-        (str(pathlib.Path("example/__main__.py")), "example", "python -m example"),
-        (str(pathlib.Path("example/cli.py")), "example", "python -m example.cli"),
-        (str(pathlib.Path("./example")), "", "example"),
-    ],
-)
-def test_detect_program_name(path, main, expected):
-    assert click.utils._detect_program_name(path, _main=MockMain(main)) == expected
-
-
-def test_expand_args(monkeypatch):
-    user = os.path.expanduser("~")
-    assert user in click.utils._expand_args(["~"])
-    monkeypatch.setenv("CLICK_TEST", "hello")
-    assert "hello" in click.utils._expand_args(["$CLICK_TEST"])
-    assert "pyproject.toml" in click.utils._expand_args(["*.toml"])
-    assert os.path.join("tests", "conftest.py") in click.utils._expand_args(
-        ["**/conftest.py"]
-    )
-    assert "*.not-found" in click.utils._expand_args(["*.not-found"])
-    # a bad glob pattern, such as a pytest identifier, should return itself
-    assert click.utils._expand_args(["test.py::test_bad"])[0] == "test.py::test_bad"
-
-
-@pytest.mark.parametrize(
-    ("value", "max_length", "expect"),
-    [
-        pytest.param("", 10, "", id="empty"),
-        pytest.param("123 567 90", 10, "123 567 90", id="equal length, no dot"),
-        pytest.param("123 567 9. aaaa bbb", 10, "123 567 9.", id="sentence < max"),
-        pytest.param("123 567\n\n 9. aaaa bbb", 10, "123 567", id="paragraph < max"),
-        pytest.param("123 567 90123.", 10, "123 567...", id="truncate"),
-        pytest.param("123 5678 xxxxxx", 10, "123...", id="length includes suffix"),
-        pytest.param(
-            "token in ~/.netrc ciao ciao",
-            20,
-            "token in ~/.netrc...",
-            id="ignore dot in word",
-        ),
-    ],
-)
-@pytest.mark.parametrize(
-    "alter",
-    [
-        pytest.param(None, id=""),
-        pytest.param(
-            lambda text: "\n\b\n" + "  ".join(text.split(" ")) + "\n", id="no-wrap mark"
-        ),
-    ],
-)
-def test_make_default_short_help(value, max_length, alter, expect):
-    assert len(expect) <= max_length
-
-    if alter:
-        value = alter(value)
-
-    out = click.utils.make_default_short_help(value, max_length)
-    assert out == expect
