+++ tests/test_utils/test__expand_args.py
@@ -1,0 +1,17 @@
+import os
+
+import click
+
+
+def test_expand_args(monkeypatch):
+    user = os.path.expanduser("~")
+    assert user in click.utils._expand_args(["~"])
+    monkeypatch.setenv("CLICK_TEST", "hello")
+    assert "hello" in click.utils._expand_args(["$CLICK_TEST"])
+    assert "pyproject.toml" in click.utils._expand_args(["*.toml"])
+    assert os.path.join("tests", "conftest.py") in click.utils._expand_args(
+        ["**/conftest.py"]
+    )
+    assert "*.not-found" in click.utils._expand_args(["*.not-found"])
+    # a bad glob pattern, such as a pytest identifier, should return itself
+    assert click.utils._expand_args(["test.py::test_bad"])[0] == "test.py::test_bad"
